맵매니저와 맵의 객체를 따로 만들어서, 스테이지 단위로 구분 완료. - 개별적인 시드 없이 고정값으로 생성되기 때문에, 스테이지가 지날때마다 시드를 기반으로 새로운 시드를 만들어주어야만 한다. - 관리가 좀더 쉽게 만듬.
