+++ Assets/Resources/Prefabs/TempDirtSet.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6806601045315322175}
-  - component: {fileID: 6806601045315322172}
+  - component: {fileID: 1284207872}
   m_Layer: 0
   m_Name: TempDirtSet
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &6806601045315322172
+--- !u!114 &1284207872
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -40,20 +40,24 @@ MonoBehaviour:
   m_GameObject: {fileID: 6806601045315322174}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7be5b1ed56aedad4599526d5bb66c987, type: 3}
+  m_Script: {fileID: 11500000, guid: e88c9fa114177514cb6e09e593fc289a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  stage_generation: 1
+  stage_startpos: 
   block_HorizLength: 40
   block_VertLength: 24
-  Seed: awsdgcxf
+  Seed: qweasfds
   dirt: {fileID: 2708950538237798945, guid: 85f506ae3eeff6f4c96f1b6480f3abe2, type: 3}
   water: {fileID: 8993033485099585298, guid: 3890692ed69304e4297527f92b893a38, type: 3}
   dummy: {fileID: 2948897419448501846, guid: 79e321601c4abf14b99bb0d42d9ed155, type: 3}
   wood: {fileID: 7393993922302359843, guid: 900d53dd3d00c1d448691acde63904e7, type: 3}
-  stone: {fileID: 3002744742036368400, guid: 8c44e2602d6a8ca4bb41c27e34e3749f, type: 3}
-  rock: {fileID: 0}
-  marker: {fileID: 0}
-  station: {fileID: 0}
+  stone: {fileID: 8998377880001305152, guid: ef398f5264c90d04986b2c1dfc84384d, type: 3}
+  rock: {fileID: 6006717006439218504, guid: 8367475a22a7a754fb0fefb2963fdecf, type: 3}
+  marker: {fileID: 4177334370399184112, guid: dc0f8b0270d48474da62b7ccd555b280, type: 3}
+  station: {fileID: 8707162244468087665, guid: 6445f745aec7ea847af5773b16c5ba8d, type: 3}
+  rail: {fileID: 4440719346928936694, guid: fb20018888818ba49aeb9185cbd45efa, type: 3}
+  lastrailpos: {fileID: 0}
   colliderBlockSet: []
   LakeLineTop: 
   LakeLineBottom:
