+++ UserSettings/EditorUserSettings.asset
@@ -29,6 +29,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c72111f192b292926237e38271427fb
       flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c7d1c1914131e293b32032e2c1336b0ae2136ebf32f
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
